requirements.txt
paddlepaddle-gpu==2.5.1.post102
paddleseg==2.8.*
pdoc